+++ src/main/java/com/ddoddo/balllog/user/dto/request/FavoriteKboTeamRequest.java
@@ -4,5 +4,5 @@ import lombok.Getter;
 
 @Getter
 public class FavoriteKboTeamRequest {
-    private Long kboTeamId;
+    private Integer kboTeamId;
 }
